# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger Employee CRUD",
        "description": "A sample API that demonstrates CRUD",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Swagger API Team"
        },
        "license": {
            "name": "MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/employees": {
            "get": {
                "summary": "Get All Employees",
                "description": "fetches and returns All the Employees",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Succesfull opertaion",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/emplpoyeesResponse"
                            }
                        }
                    },
                    "400": {}
                }
            }
        },
        "/employees/{empId}": {
            "get": {
                "summary": "Get 1 Employee",
                "description": "fetches and return 1 Employee data",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "empId",
                        "description": "employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfull opertaion",
                        "schema": {
                            "type": "object"                            
                        }
                    },
                    "400": {}
                }
            }
        }
    },
    "definitions": {
        "emplpoyeesResponse": {
            "type": "object",
            "properties": {
                "eId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sal": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                }
            }
        }
    }
}